art_gallery/requirements.txt.txt
pillow == 2.7.0
django-crispy-forms == 1.9.2
